requirements.txt
Django
djangorestframework
python-dotenv
requests
django-cors-headers
